V2

+++ web-server/src/graphql-api/mutation/delete-export-ban-list.js
@@ -1,7 +1,4 @@
-import fs from 'fs';
-import path from 'path';
-
-import { ExportBanList } from 'database/models';
+import { ExportBan, ExportBanList } from 'database/models';
 
 export default async (_, args, context) => {
   const exportBanList = await ExportBanList.findOne({
@@ -10,12 +7,15 @@ export default async (_, args, context) => {
   });
   if (exportBanList === null) throw new Error('Export ban list not found.');
 
-  await ExportBanList.deleteOne({ _id: args._id });
-
-  const exportBanListPath = path.resolve(
-    `./export-ban-lists/${exportBanList._id}.txt`
-  );
-  if (fs.existsSync(exportBanListPath)) fs.unlinkSync(exportBanListPath);
+  if (exportBanList.battlemetricsStatus === 'disabled') {
+    await ExportBanList.deleteOne({ _id: args._id });
+  } else {
+    await ExportBanList.updateOne(
+      { _id: args._id },
+      { battlemetricsStatus: 'deleted' }
+    );
+  }
+  await ExportBan.deleteMany({ banList: args._id });
 
   return exportBanList;
 };
